change main pkg name
