+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.MementoMori"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="com.malferma.mementomori.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.MementoMori">
