--- app/src/main/java/com/example/mementomori/data/const/DataPickDropDown.kt
@@ -1,213 +1,0 @@
-package com.example.mementomori.data.const
-
-data class DataPickDropDown(
-
-    val months: List<String> = listOf("01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12",),
-    var years: List<String> = listOf("1920", "1921", "1922", "1923", "1924", "1925", "1926", "1927", "1928", "1929", "1930", "1931", "1932", "1933", "1934", "1935", "1936", "1937", "1938", "1939", "1940", "1941", "1942", "1943", "1944", "1945", "1946", "1947", "1948", "1949", "1950", "1951", "1952", "1953", "1954", "1955", "1956", "1957", "1958", "1959", "1960", "1961", "1962", "1963", "1964", "1965", "1966", "1967", "1968", "1969", "1970", "1971", "1972", "1973", "1974", "1975", "1976", "1977", "1978", "1979", "1980", "1981", "1982", "1983", "1984", "1985", "1986", "1987", "1988", "1989", "1990", "1991", "1992", "1993", "1994", "1995", "1996", "1997", "1998", "1999", "2000", "2001", "2002", "2003", "2004", "2005", "2006", "2007", "2008", "2009", "2010", "2011", "2012", "2013", "2014", "2015", "2016", "2017", "2018", "2019", "2020", "2021", "2022",),
-
-    /**
-     * Dictionary with life expectancy by country as key
-     * @sample => stats for: average of both / female / male
-     * */
-    val countryDict: Map<String, List<Double>> = mapOf(
-        "Afghanistan"           to listOf(65.98, 67.59, 64.47),
-        "Albania"               to listOf(78.96, 80.48, 77.48),
-        "Algeria"               to listOf(77.50, 78.76, 76.30),
-        "Angola"                to listOf(62.22, 65.12, 59.46),
-        "Antigua and Barbuda"   to listOf(77.47, 78.58, 76.26),
-        "Argentina"             to listOf(77.17, 80.42, 73.82),
-        "Armenia"               to listOf(75.55, 78.90, 71.82),
-        "Aruba"                 to listOf(76.79, 79.10, 74.26),
-        "Australia" to listOf(83.94, 85.80, 82.08),
-        "Austria" to listOf(82.05, 84.19, 79.88),
-        "Azerbaijan" to listOf(73.33, 75.87, 70.76),
-        "Bahamas" to listOf(74.28, 76.51, 71.95),
-        "Bahrain" to listOf(77.73, 78.88, 76.87),
-        "Bangladesh" to listOf(73.57, 75.60, 71.80),
-        "Barbados" to listOf(79.64, 80.85, 78.36),
-        "Belarus" to listOf(75.20, 79.90, 70.15),
-        "Belgium" to listOf(82.17, 84.31, 80.00),
-        "Belize" to listOf(75.09, 78.25, 72.14),
-        "Benin" to listOf(62.84, 64.45, 61.23),
-        "Bhutan" to listOf(72.77, 73.33, 72.27),
-        "Bolivia" to listOf(72.35, 75.35, 69.50),
-        "Bosnia and Herzegovina" to listOf(77.93, 80.32, 75.48),
-        "Botswana" to listOf(69.86, 72.69, 66.72),
-        "Brazil" to listOf(76.57, 80.14, 73.01),
-        "Brunei" to listOf(76.35, 77.64, 75.17),
-        "Bulgaria" to listOf(75.49, 79.06, 72.05),
-        "Burkina Faso" to listOf(62.98, 63.78, 62.06),
-        "Burundi" to listOf(62.71, 64.56, 60.85),
-        "Cabo Verde" to listOf(73.58, 76.83, 70.15),
-        "Cambodia" to listOf(70.54, 72.69, 68.20),
-        "Cameroon" to listOf(60.32, 61.66, 58.99),
-        "Canada" to listOf(82.96, 84.74, 81.15),
-        "Central African Republic" to listOf(54.36, 56.58, 52.16),
-        "Chad" to listOf(55.17, 56.65, 53.73),
-        "Channel Islands" to listOf(83.60, 85.31, 81.82),
-        "Chile" to listOf(80.74, 82.80, 78.54),
-        "China" to listOf(77.47, 79.73, 75.36),
-        "Colombia" to listOf(77.87, 80.54, 75.18),
-        "Comoros" to listOf(65.03, 66.88, 63.24),
-        "Congo" to listOf(65.21, 66.73, 63.67),
-        "Costa Rica" to listOf(80.94, 83.39, 78.53),
-        "Croatia" to listOf(79.02, 82.02, 75.95),
-        "Cuba" to listOf(79.18, 81.12, 77.25),
-        "Curaçao" to listOf(79.41, 82.08, 76.42),
-        "Cyprus" to listOf(81.51, 83.45, 79.55),
-        "Czech Republic (Czechia)" to listOf(79.85, 82.35, 77.33),
-        "Côte d'Ivoire" to listOf(58.75, 60.13, 57.50),
-        "Denmark" to listOf(81.40, 83.27, 79.54),
-        "Djibouti" to listOf(67.87, 70.16, 65.84),
-        "Dominican Republic" to listOf(74.65, 77.90, 71.58),
-        "DR Congo" to listOf(61.60, 63.21, 60.01),
-        "Ecuador" to listOf(77.71, 80.45, 75.05),
-        "Egypt" to listOf(72.54, 74.95, 70.23),
-        "El Salvador" to listOf(74.06, 78.48, 69.27),
-        "Equatorial Guinea" to listOf(59.82, 61.08, 58.76),
-        "Eritrea" to listOf(67.48, 69.75, 65.26),
-        "Estonia" to listOf(79.18, 83.06, 74.98),
-        "Eswatini" to listOf(61.05, 65.67, 56.98),
-        "Ethiopia" to listOf(67.81, 69.80, 65.86),
-        "Fiji" to listOf(67.91, 69.97, 66.00),
-        "Finland" to listOf(82.48, 85.14, 79.82),
-        "France" to listOf(83.13, 85.82, 80.32),
-        "French Guiana" to listOf(80.53, 83.38, 77.80),
-        "French Polynesia" to listOf(78.23, 80.41, 76.23),
-        "Gabon" to listOf(67.03, 69.27, 64.93),
-        "Gambia" to listOf(63.26, 64.73, 61.80),
-        "Georgia" to listOf(74.24, 78.54, 69.85),
-        "Germany" to listOf(81.88, 84.14, 79.62),
-        "Ghana" to listOf(64.94, 66.13, 63.78),
-        "Greece" to listOf(82.80, 85.08, 80.52),
-        "Grenada" to listOf(72.59, 75.16, 70.24),
-        "Guadeloupe" to listOf(82.74, 85.94, 79.16),
-        "Guam" to listOf(80.74, 83.98, 77.63),
-        "Guatemala" to listOf(75.05, 77.89, 72.14),
-        "Guinea" to listOf(62.64, 63.25, 61.79),
-        "Guinea-Bissau" to listOf(59.38, 61.33, 57.31),
-        "Guyana" to listOf(70.26, 73.53, 67.22),
-        "Haiti" to listOf(64.99, 67.21, 62.77),
-        "Honduras" to listOf(75.87, 78.14, 73.57),
-        "Hong Kong" to listOf(85.29, 88.17, 82.38),
-        "Hungary" to listOf(77.31, 80.66, 73.78),
-        "Iceland" to listOf(83.52, 84.90, 82.15),
-        "India" to listOf(70.42, 71.80, 69.16),
-        "Indonesia" to listOf(72.32, 74.64, 70.12),
-        "Iran" to listOf(77.33, 78.54, 76.22),
-        "Iraq" to listOf(71.08, 73.21, 68.96),
-        "Ireland" to listOf(82.81, 84.32, 81.29),
-        "Israel" to listOf(83.49, 84.91, 81.98),
-        "Israel" to listOf(83.49, 84.91, 81.98),
-        "Italy" to listOf(84.01, 85.97, 81.90),
-        "Jamaica" to listOf(74.88, 76.60, 73.17),
-        "Japan" to listOf(85.03, 88.09, 81.91),
-        "Jordan" to listOf(75.01, 76.82, 73.28),
-        "Kazakhstan" to listOf(73.90, 77.97, 69.55),
-        "Kenya" to listOf(67.47, 69.87, 65.04),
-        "Kiribati" to listOf(69.17, 73.12, 65.00),
-        "Kuwait" to listOf(75.85, 77.06, 75.09),
-        "Kyrgyzstan" to listOf(71.95, 76.03, 67.85),
-        "Laos" to listOf(68.89, 70.79, 67.04),
-        "Latvia" to listOf(75.73, 80.37, 70.81),
-        "Lebanon" to listOf(79.27, 81.17, 77.53),
-        "Lesotho" to listOf(55.65, 58.90, 52.52),
-        "Liberia" to listOf(65.00, 66.48, 63.54),
-        "Libya" to listOf(73.44, 76.46, 70.61),
-        "Lithuania" to listOf(76.41, 81.70, 70.97),
-        "Luxembourg" to listOf(82.79, 84.76, 80.83),
-        "Macao" to listOf(84.68, 87.62, 81.73),
-        "Madagascar" to listOf(68.21, 69.92, 66.53),
-        "Malawi" to listOf(65.62, 68.77, 62.45),
-        "Malaysia" to listOf(76.65, 78.78, 74.71),
-        "Maldives" to listOf(79.89, 81.58, 78.53),
-        "Mali" to listOf(60.54, 61.39, 59.69),
-        "Malta" to listOf(83.06, 84.68, 81.37),
-        "Martinique" to listOf(83.13, 86.10, 79.85),
-        "Mauritania" to listOf(62.08, 63.45, 60.90),
-        "Mauritius" to listOf(75.67, 77.37, 74.01),
-        "Mexico" to listOf(79.17, 80.88, 77.56),
-        "Moldova" to listOf(71.12, 75.50, 66.73),
-        "Monaco" to listOf(83.84, 85.91, 81.71),
-        "Mongolia" to listOf(68.06, 70.12, 66.04),
-        "Montenegro" to listOf(78.92, 81.34, 76.65),
-        "Morocco" to listOf(76.27, 78.90, 73.78),
-        "Mozambique" to listOf(69.66, 71.17, 68.22),
-        "Myanmar" to listOf(68.06, 70.79, 65.42),
-        "Namibia" to listOf(74.64, 77.34, 71.92),
-        "Nepal" to listOf(69.63, 71.34, 67.89),
-        "Netherlands" to listOf(82.12, 84.06, 80.14),
-        "New Caledonia" to listOf(79.71, 82.71, 76.80),
-        "New Zealand" to listOf(82.86, 84.60, 81.09),
-        "Nicaragua" to listOf(76.71, 79.72, 73.65),
-        "Niger" to listOf(60.82, 62.18, 59.67),
-        "Nigeria" to listOf(56.44, 57.79, 55.27),
-        "North Macedonia" to listOf(78.67, 81.01, 76.28),
-        "Norway" to listOf(82.92, 84.78, 81.05),
-        "Oman" to listOf(78.29, 80.33, 76.27),
-        "Pakistan" to listOf(66.46, 68.47, 64.45),
-        "Panama" to listOf(78.03, 80.46, 75.62),
-        "Papua New Guinea" to listOf(66.71, 68.23, 65.18),
-        "Paraguay" to listOf(75.84, 78.45, 73.28),
-        "Peru" to listOf(76.98, 79.38, 74.59),
-        "Philippines" to listOf(73.01, 75.34, 70.68),
-        "Poland" to listOf(80.24, 82.86, 77.59),
-        "Portugal" to listOf(81.63, 83.92, 79.27),
-        "Puerto Rico" to listOf(80.88, 83.13, 78.56),
-        "Qatar" to listOf(80.78, 82.54, 79.03),
-        "Romania" to listOf(78.69, 81.41, 76.07),
-        "Russia" to listOf(73.54, 77.37, 69.66),
-        "Rwanda" to listOf(69.12, 70.97, 67.36),
-        "Saint Kitts and Nevis" to listOf(78.45, 81.39, 75.48),
-        "Saint Lucia" to listOf(78.61, 81.14, 76.06),
-        "Saint Martin" to listOf(81.72, 85.43, 77.86),
-        "Samoa" to listOf(73.75, 75.97, 71.69),
-        "Saudi Arabia" to listOf(75.69,77.37, 74.47),
-        "Senegal" to listOf(68.87, 70.88, 66.64),
-        "Serbia" to listOf(76.47, 79.05, 73.89),
-        "Seychelles" to listOf(73.74, 77.71, 70.26),
-        "Sierra Leone" to listOf(55.92, 56.78, 55.01),
-        "Singapore" to listOf(84.07, 86.15, 82.06),
-        "Slovakia" to listOf(78.00, 81.35, 74.59),
-        "Slovenia" to listOf(81.85, 84.44, 79.26),
-        "Solomon Islands" to listOf(73.38, 75.31, 71.60),
-        "Somalia" to listOf(58.34, 60.11, 56.62),
-        "South Africa" to listOf(64.88, 68.42, 61.46),
-        "South Korea" to listOf(83.50, 86.42, 80.46),
-        "South Sudan" to listOf(58.74, 60.31, 57.21),
-        "Spain" to listOf(83.99, 86.68, 81.27),
-        "Sri Lanka" to listOf(77.56, 80.74, 74.25),
-        "State of Palestine" to listOf(74.62, 76.38, 72.92),
-        "Sudan" to listOf(66.09, 68.03, 64.18),
-        "Suriname" to listOf(72.13, 75.55, 68.88),
-        "Sweden" to listOf(83.33, 84.97, 81.69),
-        "Switzerland" to listOf(84.25, 86.02, 82.42),
-        "Syria" to listOf(76.06, 79.11, 73.13),
-        "Taiwan" to listOf(81.04, 83.64, 78.49),
-        "Tajikistan" to listOf(71.76, 74.11, 69.50),
-        "Tanzania" to listOf(66.39, 68.25, 64.52),
-        "Thailand" to listOf(77.74, 81.34, 74.16),
-        "Timor-Leste" to listOf(70.18, 72.36, 68.14),
-        "Togo" to listOf(62.13, 63.08, 61.16),
-        "Tonga" to listOf(71.32, 73.38, 69.28),
-        "Trinidad and Tobago" to listOf(73.91, 76.66, 71.24),
-        "Tunisia" to listOf(77.36, 79.34, 75.37),
-        "Turkey" to listOf(78.45, 81.21, 75.57),
-        "Turkmenistan" to listOf(68.63, 72.19, 65.10),
-        "Uganda" to listOf(64.38, 66.67, 61.99),
-        "Ukraine" to listOf(72.50, 77.27, 67.56),
-        "United Arab Emirates" to listOf(78.46, 79.80, 77.79),
-        "United Kingdom" to listOf(81.77, 83.28, 80.22),
-        "United States" to listOf(79.11, 81.65, 76.61),
-        "Uruguay" to listOf(78.43, 81.88, 74.75),
-        "Uzbekistan" to listOf(72.04, 74.26, 69.80),
-        "Vanuatu" to listOf(70.99, 72.75, 69.40),
-        "Venezuela" to listOf(72.34, 76.25, 68.59),
-        "Vietnam" to listOf(75.77, 79.85, 71.73),
-        "Western Sahara" to listOf(71.08, 73.19, 69.36),
-        "Yemen" to listOf(66.44, 68.16, 64.72),
-        "Zambia" to listOf(64.70, 67.71, 61.66),
-        "Zimbabwe" to listOf(62.16, 63.66, 60.39),
-    )
-)
-
