--- app/src/main/java/com/example/mementomori/data/lastInput/LastInputDao.kt
@@ -1,17 +1,0 @@
-package com.example.mementomori.data.lastInput
-
-import androidx.lifecycle.LiveData
-import androidx.room.*
-
-@Dao
-interface LastInputDao {
-
-    @Insert(onConflict = OnConflictStrategy.IGNORE)
-    suspend fun addLastInput(lastInput: LastInput)
-
-    @Query("SELECT * FROM LastInput")
-    fun readAllData(): LiveData<List<LastInput>>
-
-    @Update
-    fun updateLastInput(entity: LastInput)
-}
