--- app/src/main/java/com/example/mementomori/screen/mainChart/MainChartViewModel.kt
@@ -1,11 +1,0 @@
-package com.example.mementomori.screen.mainChart
-
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.setValue
-import androidx.lifecycle.ViewModel
-
-class MainChartViewModel: ViewModel() {
-
-
-}
