--- app/src/main/java/com/example/mementomori/ui/theme/Color.kt
@@ -1,70 +1,0 @@
-package com.example.mementomori.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-val none = Color(0xFF4A1DB3)
-
-val Purple80 = Color(0xFFD0BCFF)
-val PurpleGrey80 = Color(0xFFCCC2DC)
-val Pink80 = Color(0xFFEFB8C8)
-
-val Purple40 = Color(0xFF6650a4)
-val PurpleGrey40 = Color(0xFF625b71)
-val Pink40 = Color(0xFF7D5260)
-
-val Teal200 = Color(0xFF03DAC5)
-val Teal300 = Color(0xFF00B7FF)
-
-/**
- * main
- * */
-val main_fff = Color(0xFFFFFFFF)
-val main_100 = Color(0xFFECF0F3)
-val main_200 = Color(0xFFDDE5F0)
-val main_300 = Color(0xFFD1D9E6)
-val main_400 = Color(0xFF97A7C3)
-val main_500 = Color(0xFF284577)
-val main_600 = Color(0xFF193566)
-val main_700 = Color(0xFF212C3C)
-val main_800 = Color(0xFF111A2E)
-val main_900 = Color(0xFF0E111A)
-val main_000 = Color(0xFF000000)
-var modalTransBground = Color(0xFF575757)
-
-/**
- * Custom colors
- * */
-val chart_1 = Color(0xFFE9E8E8)
-val chart_2 = Color(0xFF2EFF27)
-val chart_3 = Color(0xFF00CCBB)
-val chart_4 = Color.Yellow
-val chart_5 = Color(0xFFCACACA)
-val chart_6 = Color(0xFF00C7B0)
-//val chart_x = Color(0xFFC73F00)
-val chart_x = Color(0xFFF03202)
-val chart_bg_above = Color(0xFF00FFE2)
-val chart_above = Color(0xFF007501)
-//val chart_x = Color(0xFF525252)
-val bgHeader = Color(0xFF757575)
-val headerItems = Color(0xFFF1F1F1)
-
-/**
- * Font colors light
- * */
-val font_100l = Color(0xFF43649B)
-val font_200l = Color(0xFFFFFFFF)
-val font_300l = Color(0xFFFFFFFF)
-
-/**
- * Font colors dark
- * */
-val font_100d = Color(0xFFFFFFFF)
-val font_200d = Color(0xFF000000)
-val font_300d = Color(0xFF000000)
-
-/**
- * Bottom nav bar
- * */
-// Light
-
-// Dark
