--- app/src/main/java/com/example/mementomori/ui/theme/CustomeTheme.kt
@@ -1,10 +1,0 @@
-package com.example.mementomori.ui.theme
-
-import androidx.compose.material.MaterialTheme
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.ReadOnlyComposable
-
-val MaterialTheme.myColors: CustomColors
-    @Composable
-    @ReadOnlyComposable
-    get() = LocalColors.current
