--- app/src/main/java/com/example/mementomori/ui/theme/Type.kt
@@ -1,121 +1,0 @@
-package com.example.mementomori.ui.theme
-
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Typography
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.em
-import androidx.compose.ui.unit.sp
-
-// Set of Material typography styles to start with
-val Typography = Typography(
-    h2 = TextStyle(
-        fontFamily = FontFamily.Default,
-        fontSize = 22.sp,
-        fontWeight = FontWeight.Medium,
-    ),
-    h3 = TextStyle(
-        fontFamily = FontFamily.Default,
-        fontSize = 18.sp,
-        fontWeight = FontWeight.Medium,
-    ),
-    h4 = TextStyle(
-        fontFamily = FontFamily.Default,
-        fontSize = 16.sp,
-        fontWeight = FontWeight.Medium,
-    ),
-    h5 = TextStyle(
-        fontFamily = FontFamily.Default,
-        fontSize = 15.sp,
-        fontWeight = FontWeight.Normal,
-    ),
-    body1 = TextStyle(
-        fontFamily = FontFamily.Default,
-        fontWeight = FontWeight.Normal,
-        fontSize = 16.sp
-    )
-    /* Other default text styles to override
-    button = TextStyle(
-        fontFamily = FontFamily.Default,
-        fontWeight = FontWeight.W500,
-        fontSize = 14.sp
-    ),
-    caption = TextStyle(
-        fontFamily = FontFamily.Default,
-        fontWeight = FontWeight.Normal,
-        fontSize = 12.sp
-    )
-    */
-)
-
-val Typography.UserInput: TextStyle
-    @Composable
-    get() {
-        return  TextStyle(
-            fontFamily = FontFamily.Default,
-            fontWeight = FontWeight.Bold,
-            fontSize = 20.sp,
-            color = MaterialTheme.myColors.bgHeader,
-        )
-    }
-
-val Typography.Checkbox: TextStyle
-    @Composable
-    get() {
-        return  TextStyle(
-            fontFamily = FontFamily.Default,
-            fontWeight = FontWeight.Normal,
-            fontSize = 16.sp,
-        )
-    }
-
-val Typography.HeaderContainer: TextStyle
-    @Composable
-    get() {
-        return  TextStyle(
-            fontFamily = FontFamily.Default,
-            fontWeight = FontWeight.Medium,
-            fontSize = 15.sp,
-            textAlign = TextAlign.Center,
-            color = MaterialTheme.myColors.headerItems,
-        )
-    }
-
-val Typography.HeaderContainerInfo: TextStyle
-    @Composable
-    get() {
-        return  TextStyle(
-            fontFamily = FontFamily.Default,
-            fontWeight = FontWeight.Medium,
-            fontSize = 15.sp,
-            textAlign = TextAlign.Left,
-            color = MaterialTheme.myColors.headerItems,
-        )
-    }
-
-val Typography.AgeGroup: TextStyle
-    @Composable
-    get() {
-        return  TextStyle(
-            fontFamily = FontFamily.Default,
-            fontWeight = FontWeight.Bold,
-            fontSize = 30.sp,
-            textAlign = TextAlign.Center,
-            color = MaterialTheme.myColors.bgHeader,
-        )
-    }
-
-val Typography.AgeGroupMonths: TextStyle
-    @Composable
-    get() {
-        return  TextStyle(
-            fontFamily = FontFamily.Default,
-            fontWeight = FontWeight.Bold,
-            fontSize = 20.sp,
-            textAlign = TextAlign.Center,
-            color = MaterialTheme.myColors.bgHeader,
-        )
-    }
