+++ app/src/main/java/com/malferma/mementomori/data/lastInput/LastInputDao.kt
@@ -1,0 +1,17 @@
+package com.malferma.mementomori.data.lastInput
+
+import androidx.lifecycle.LiveData
+import androidx.room.*
+
+@Dao
+interface LastInputDao {
+
+    @Insert(onConflict = OnConflictStrategy.IGNORE)
+    suspend fun addLastInput(lastInput: LastInput)
+
+    @Query("SELECT * FROM LastInput")
+    fun readAllData(): LiveData<List<LastInput>>
+
+    @Update
+    fun updateLastInput(entity: LastInput)
+}
