+++ app/src/main/java/com/malferma/mementomori/screen/mainChart/MainChartViewModel.kt
@@ -1,0 +1,8 @@
+package com.malferma.mementomori.screen.mainChart
+
+import androidx.lifecycle.ViewModel
+
+class MainChartViewModel: ViewModel() {
+
+
+}
