+++ app/src/main/java/com/malferma/mementomori/ui/theme/CustomeTheme.kt
@@ -1,0 +1,10 @@
+package com.malferma.mementomori.ui.theme
+
+import androidx.compose.material.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.ReadOnlyComposable
+
+val MaterialTheme.myColors: CustomColors
+    @Composable
+    @ReadOnlyComposable
+    get() = LocalColors.current
